requirements.txt
Django~=4.1.2
djangorestframework~=3.14.0
djangorestframework_simplejwt~=5.2.2
pytest_django~=4.5.2
pytest_factoryboy~=2.5.0
psycopg2_binary~=2.9.4
pytest~=7.1.3